+++ src/command_modules/azure-cli-lab/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+unreleased
++++++++++++++++++++++
+
+* Adding show, delete & list commands for environment in the lab.
 
 0.0.2 (2017-04-17)
 +++++++++++++++++++++